Improve local development workflow
- Remove graphql-yoga
- Add skaffold for local development
- Add docker-compose for server container so I can run commands within a
container, and sync those changes out to the local filesystem.  An
example would be running something like `npm init` or `npm install
express`, and have those changes be synced to the local filesystem, so
they can be commited to version control, and then later added to the
docker image.
- Improve Readme
- Get apollo server express, express, typescript, barebones graphql and
more